airflow/requirements.txt
apache-airflow==2.7.2
psycopg2-binary==2.9.7
requests==2.31.0
pandas==2.2.2